+++ package.json
@@ -10,6 +10,10 @@
     "test": "vitest run --coverage"
   },
   "dependencies": {
+    "@chakra-ui/react": "^2.8.2",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "framer-motion": "^11.2.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
