+++ src/components/Footer.jsx
@@ -1,0 +1,7 @@
+export function Footer() {
+    return (
+      <footer>
+        <p>&copy; 2023 Your Ecommerce Store. All rights reserved.</p>
+      </footer>
+    );
+  }
