+++ src/components/ProductSection.jsx
@@ -1,0 +1,40 @@
+import Button from "./common/Button";
+
+export function ProductSection() {
+    return (
+      <section className="product-section">
+        <div className="product-section-container">
+          <h2 className="product-featured-heading">Featured Products</h2>
+          <div className="product-card-list">
+            <div className="product-card">
+              <img
+                src="https://placehold.co/300x300"
+                alt="The Lorem Ipsum for photos."
+              />
+              <h3>Premium Bluetooth Headphones</h3>
+              <p className="product-price">2,400 Baht</p>
+              <Button buttonText="Add to Cart" buttonStyle="add-to-cart-button" />
+            </div>
+            <div className="product-card">
+              <img
+                src="https://placehold.co/300x300"
+                alt="The Lorem Ipsum for photos."
+              />
+              <h3>Classic Leather Watch</h3>
+              <p className="product-price">4,000 Baht</p>
+              <Button buttonText="Add to Cart" buttonStyle="add-to-cart-button" />
+            </div>
+            <div className="product-card">
+              <img
+                src="https://placehold.co/300x300"
+                alt="The Lorem Ipsum for photos."
+              />
+              <h3>Organic Green Tea</h3>
+              <p className="product-price">79.99 Baht</p>
+              <Button buttonText="Add to Cart" buttonStyle="add-to-cart-button" />
+            </div>
+          </div>
+        </div>
+      </section>
+    );
+  }
