+++ src/components/common/Button.jsx
@@ -1,0 +1,5 @@
+function Button(props) {
+    return <button className={props.buttonStyle}>{props.buttonText}</button>;
+  }
+
+  export default Button;
